tests/wpt/tests/html/canvas/element/line-styles/2d.line.width.scaledefault.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.line.width.scaledefault</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.line.width.scaledefault</h1>
<p class="desc">Default lineWidth strokes are affected by scale transformations</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Default lineWidth strokes are affected by scale transformations");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);

  ctx.scale(50, 50);
  ctx.strokeStyle = '#0f0';
  ctx.moveTo(0, 0.5);
  ctx.lineTo(2, 0.5);
  ctx.stroke();

  _assertPixel(canvas, 25,25, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 75,25, 0,255,0,255);
  _assertPixel(canvas, 50,5, 0,255,0,255);
  _assertPixel(canvas, 50,45, 0,255,0,255);

});
</script>
