LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/line-styles/2d.line.width.scaledefault-expected.txt

2d.line.width.scaledefault
Default lineWidth strokes are affected by scale transformations
Actual output:
Expected output:


FAIL Default lineWidth strokes are affected by scale transformations assert_equals: Red channel of the pixel at (25, 25) expected 0 but got 128
